moved function into modal file, to fix dependency cycle
